+++ test/runner.js
@@ -8,6 +8,7 @@ var advanced = require('./specs/advanced');
 var transition = require('./specs/transition');
 var errorHandling = require('./specs/error-handling');
 var transitionLifecycle = require('./specs/transition-lifecycle');
+var returnValue = require('./specs/return-value');
 
 var promises = {
   Default: defaultPromise,
@@ -27,5 +28,6 @@ Object.keys(promises).forEach(function (promise) {
     transition(promises[promise]);
     errorHandling(promises[promise]);
     transitionLifecycle(promises[promise]);
+    returnValue(promises[promise]);
   })
 });